requirements.txt
Flask>=2.0
pytest>=6.0
requests>=2.0
beautifulsoup4>=4.9
openpyxl>=3.0
psycopg2-binary>=2.9
flask-cors>=4.0
gunicorn
psycopg2-binary

